auth finished

+++ app/Http/Controllers/ListingController.php
@@ -51,6 +51,12 @@ class ListingController extends Controller
     }
 
     public function update(Request $request, Listing $listing) {
+
+        //Make sure logged in user is owner
+        if($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
+
         $formFields = $request->validate([
             'title' => 'required',
             'company' => 'required',
@@ -71,10 +77,14 @@ class ListingController extends Controller
     }
 
     public function destroy(Listing $listing) {
+
+        if($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
+
         $listing->delete();
         return redirect('/')->with('message', 'Listing deleted successfully');
     }
-
     //Manage listings
     public function manage() {
         return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
